pyproject.toml
[tool.poetry]
name = "employee-directory"
version = "0.1.0"
description = ""
authors = ["el-collins <charlescollins476@gmail.com>"]
readme = "README.md"
packages = [{include = "employee_directory"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.104.1"
uvicorn = "0.24.0"
sqlalchemy = "2.0.23"
pydantic = "2.5.2"
pytest = "7.4.3"
python-dotenv = "1.0.0"
psycopg2-binary = "2.9.9"
httpx = "0.25.2"
sqlmodel = "^0.0.22"
pymysql = "^1.1.1"
alembic = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
